+++ src/builder/pod/mod.rs
@@ -3,6 +3,7 @@ pub mod security;
 pub mod volume;
 
 use crate::builder::meta::ObjectMetaBuilder;
+use crate::commons::product_image_selection::ResolvedProductImage;
 use crate::error::{Error, OperatorResult};
 
 use k8s_openapi::{
@@ -287,6 +288,19 @@ impl PodBuilder {
         self.image_pull_secrets
             .get_or_insert_with(Vec::new)
             .extend(secrets.map(|s| LocalObjectReference { name: Some(s) }));
+        self
+    }
+
+    /// Extend the pod's image_pull_secrets field with the pull secrets from a given [ResolvedProductImage]
+    pub fn image_pull_secrets_from_product_image(
+        &mut self,
+        product_image: &ResolvedProductImage,
+    ) -> &mut Self {
+        if let Some(pull_secrets) = &product_image.pull_secrets {
+            self.image_pull_secrets
+                .get_or_insert_with(Vec::new)
+                .extend_from_slice(pull_secrets);
+        }
         self
     }