+++ src/labels.rs
@@ -5,6 +5,8 @@ use std::collections::BTreeMap;
 
 #[cfg(doc)]
 use crate::builder::ObjectMetaBuilder;
+#[cfg(doc)]
+use crate::commons::product_image_selection::ResolvedProductImage;
 
 const APP_KUBERNETES_LABEL_BASE: &str = "app.kubernetes.io/";
 
@@ -31,10 +33,16 @@ pub struct ObjectLabels<'a, T> {
     pub owner: &'a T,
     /// The name of the app being managed, such as `zookeeper`
     pub app_name: &'a str,
-    /// The version of the app being managed (not of the operator)
+    /// The version of the app being managed (not of the operator).
     ///
-    /// This version should include the Stackable version, such as `0.1.0-stackable0.1.0`. In the case of custom product images,
-    /// the tag should be appended, such as `<productversion>-<tag>`. However, this is pure documentation and should not be parsed.
+    /// If setting this label on a Stackable product then please use [`ResolvedProductImage::app_version_label`]
+    ///
+    /// This version should include the Stackable version, such as `3.0.0-stackable0.1.0`.
+    /// If the Stackable version is not known, then the product version should be used together with a suffix (if possible).
+    /// If a custom product image is provided by the user (in which case only the product version is known),
+    /// then the format `3.0.0-<tag-of-custom-image>` should be used.
+    ///
+    /// However, this is pure documentation and should not be parsed.
     pub app_version: &'a str,
     /// The DNS-style name of the operator managing the object (such as `zookeeper.stackable.tech`)
     pub operator_name: &'a str,